refactor(sdk,cli): move getPluginDisplayName to shared storage
The TUI-local copy becomes the single shared derivation so the session
config layer and the settings tab resolve the same plugin names.
